Center Continue button
