+++ package.json
@@ -25,6 +25,7 @@
     "d3-ease": "^2.0.0",
     "express": "^4.17.1",
     "framer-motion": "^2.6.5",
+    "html2canvas": "^1.0.0-rc.7",
     "mapbox-gl": "^1.11.1",
     "next": "^9.5.2",
     "next-i18next": "^4.5.0",
