+++ src/features/public/Donations/components/TreeDonation.tsx
@@ -4,7 +4,6 @@ import Sugar from 'sugar';
 import { getCountryDataBy } from '../../../../utils/countryUtils';
 import { formatAmountForStripe } from '../../../../utils/stripeHelpers';
 import PaymentProgress from '../../../common/ContentLoaders/Donations/PaymentProgress';
-import AnimatedButton from '../../../common/InputTypes/AnimatedButton';
 import ToggleSwitch from '../../../common/InputTypes/ToggleSwitch';
 import { TreeDonationProps } from '../../../common/types/donations';
 import SelectCurrencyModal from '../components/treeDonation/SelectCurrencyModal';
@@ -44,60 +43,7 @@ function TreeDonation({
   );
   const [paymentError, setPaymentError] = React.useState('');
 
-  const stripeAllowedCountries = [
-    'AE',
-    'AT',
-    'AU',
-    'BE',
-    'BG',
-    'BR',
-    'CA',
-    'CH',
-    'CI',
-    'CR',
-    'CY',
-    'CZ',
-    'DE',
-    'DK',
-    'DO',
-    'EE',
-    'ES',
-    'FI',
-    'FR',
-    'GB',
-    'GR',
-    'GT',
-    'HK',
-    'HU',
-    'ID',
-    'IE',
-    'IN',
-    'IT',
-    'JP',
-    'LT',
-    'LU',
-    'LV',
-    'MT',
-    'MX',
-    'MY',
-    'NL',
-    'NO',
-    'NZ',
-    'PE',
-    'PH',
-    'PL',
-    'PT',
-    'RO',
-    'SE',
-    'SG',
-    'SI',
-    'SK',
-    'SN',
-    'TH',
-    'TT',
-    'US',
-    'UY',
-  ];
+
   const [isPaymentProcessing, setIsPaymentProcessing] = React.useState(false);
   const taxDeductSwitchOn = () => {
     setIsTaxDeductible(!isTaxDeductible);
@@ -165,214 +111,202 @@ function TreeDonation({
   return isPaymentProcessing ? (
     <PaymentProgress isPaymentProcessing={isPaymentProcessing} />
   ) : (
-    <>
-      <div
-        className={styles.cardContainer}
-        style={{ alignSelf: isGift ? 'start' : 'center' }}
-      >
-        <div className={styles.header}>
-          <div onClick={onClose} className={styles.headerCloseIcon}>
-            <Close color={styles.primaryFontColor} />
-          </div>
-          <div className={styles.headerTitle}>Tree Donation</div>
-        </div>
-
-        <div className={styles.plantProjectName}>
-          To {project.name} by {project.tpo.name}
-        </div>
-
+      <>
         <div
-          className={
-            isTaxDeductible ? styles.currencyRateDisabled : styles.currencyRate
-          }
-          onClick={
-            !isTaxDeductible ? () => setOpenCurrencyModal(true) : () => {}
-          }
+          className={styles.cardContainer}
+          style={{ alignSelf: isGift ? 'start' : 'center' }}
         >
+          <div className={styles.header}>
+            <div onClick={onClose} className={styles.headerCloseIcon}>
+              <Close color={styles.primaryFontColor} />
+            </div>
+            <div className={styles.headerTitle}>Tree Donation</div>
+          </div>
+
+          <div className={styles.plantProjectName}>
+            To {project.name} by {project.tpo.name}
+          </div>
+
           <div
             className={
-              isTaxDeductible ? styles.currencyDisabled : styles.currency
+              isTaxDeductible ? styles.currencyRateDisabled : styles.currencyRate
+            }
+            onClick={
+              !isTaxDeductible ? () => setOpenCurrencyModal(true) : () => { }
             }
           >
-            {currency}
+            <div
+              className={
+                isTaxDeductible ? styles.currencyDisabled : styles.currency
+              }
+            >
+              {currency}
+            </div>
+            <div className={styles.downArrow}>
+              <DownArrow color={isTaxDeductible ? 'grey' : '#87B738'} />
+            </div>
+            <div className={styles.rate}>
+              {Number(treeCost).toFixed(2)} per tree
           </div>
-          <div className={styles.downArrow}>
-            <DownArrow color={isTaxDeductible ? 'grey' : '#87B738'} />
           </div>
-          <div className={styles.rate}>
-            {Number(treeCost).toFixed(2)} per tree
-          </div>
-        </div>
 
-        <div className={styles.isGiftDonation}>
-          <div className={styles.isGiftDonationText}>
-            My donation is a gift to someone
+          <div className={styles.isGiftDonation}>
+            <div className={styles.isGiftDonationText}>
+              My donation is a gift to someone
           </div>
-          <ToggleSwitch
-            checked={isGift}
-            onChange={() => setIsGift(!isGift)}
-            name="checkedA"
-            inputProps={{ 'aria-label': 'secondary checkbox' }}
-          />
-        </div>
+            <ToggleSwitch
+              checked={isGift}
+              onChange={() => setIsGift(!isGift)}
+              name="checkedA"
+              inputProps={{ 'aria-label': 'secondary checkbox' }}
+            />
+          </div>
 
-        {isGift ? (
-          <GiftForm
-            isGift={isGift}
-            giftDetails={giftDetails}
-            setGiftDetails={setGiftDetails}
-          />
-        ) : null}
+          {isGift ? (
+            <GiftForm
+              isGift={isGift}
+              giftDetails={giftDetails}
+              setGiftDetails={setGiftDetails}
+            />
+          ) : null}
 
-        <div className={styles.selectTreeCount}>
-          {treeCountOptions.map((option) => (
+          <div className={styles.selectTreeCount}>
+            {treeCountOptions.map((option) => (
+              <motion.div
+                whileHover={{ scale: 1.05 }}
+                whileTap={{ scale: 0.98 }}
+                onClick={() => setTreeCount(option)}
+                key={option}
+                className={
+                  treeCount === option
+                    ? styles.treeCountOptionSelected
+                    : styles.treeCountOption
+                }
+              >
+                <div className={styles.treeCountOptionTrees}>{option}</div>
+                <div className={styles.treeCountOptionTrees}>Trees</div>
+              </motion.div>
+            ))}
             <motion.div
               whileHover={{ scale: 1.05 }}
               whileTap={{ scale: 0.98 }}
-              onClick={() => setTreeCount(option)}
-              key={option}
-              className={
-                treeCount === option
-                  ? styles.treeCountOptionSelected
-                  : styles.treeCountOption
-              }
+              className={styles.treeCountOption}
+              style={{ minWidth: '65%', flexDirection: 'row' }}
             >
-              <div className={styles.treeCountOptionTrees}>{option}</div>
+              <input
+                className={styles.customTreeInput}
+                onInput={(e) => {
+                  // replaces any character other than number to blank
+                  e.target.value = e.target.value.replace(/[^0-9]/g, '');
+
+                  //  if length of input more than 12, display only 12 digits
+                  if (e.target.value.toString().length >= 12) {
+                    e.target.value = e.target.value.toString().slice(0, 12);
+                  }
+                }}
+                type="text"
+                onChange={(e) => setCustomTreeValue(e)}
+              />
               <div className={styles.treeCountOptionTrees}>Trees</div>
             </motion.div>
-          ))}
-          <motion.div
-            whileHover={{ scale: 1.05 }}
-            whileTap={{ scale: 0.98 }}
-            className={styles.treeCountOption}
-            style={{ minWidth: '65%', flexDirection: 'row' }}
-          >
-            <input
-              className={styles.customTreeInput}
-              onInput={(e) => {
-                // replaces any character other than number to blank
-                e.target.value = e.target.value.replace(/[^0-9]/g, '');
+          </div>
 
-                //  if length of input more than 12, display only 12 digits
-                if (e.target.value.toString().length >= 12) {
-                  e.target.value = e.target.value.toString().slice(0, 12);
-                }
-              }}
-              type="text"
-              onChange={(e) => setCustomTreeValue(e)}
-            />
-            <div className={styles.treeCountOptionTrees}>Trees</div>
-          </motion.div>
-        </div>
-
-        {project.taxDeductionCountries.length > 0 && (
-          <React.Fragment>
-            <div className={styles.isTaxDeductible}>
-              <div className={styles.isTaxDeductibleText}>
-                Send me a tax deduction receipt for
+          {project.taxDeductionCountries.length > 0 && (
+            <React.Fragment>
+              <div className={styles.isTaxDeductible}>
+                <div className={styles.isTaxDeductibleText}>
+                  Send me a tax deduction receipt for
               </div>
-              <ToggleSwitch
-                checked={isTaxDeductible}
-                onChange={taxDeductSwitchOn}
-                name="checkedB"
-                inputProps={{ 'aria-label': 'secondary checkbox' }}
-              />
-            </div>
+                <ToggleSwitch
+                  checked={isTaxDeductible}
+                  onChange={taxDeductSwitchOn}
+                  name="checkedB"
+                  inputProps={{ 'aria-label': 'secondary checkbox' }}
+                />
+              </div>
 
-            <div
-              className={
-                isTaxDeductible
-                  ? styles.taxDeductible
-                  : styles.taxDeductibleDisabled
-              }
-              onClick={
-                isTaxDeductible
-                  ? () => setOpenTaxDeductionModal(true)
-                  : () => {}
-              }
-            >
               <div
                 className={
                   isTaxDeductible
-                    ? styles.taxDeductibleCountry
-                    : styles.taxDeductibleCountryDisabled
+                    ? styles.taxDeductible
+                    : styles.taxDeductibleDisabled
+                }
+                onClick={
+                  isTaxDeductible
+                    ? () => setOpenTaxDeductionModal(true)
+                    : () => { }
                 }
               >
-                {project.taxDeductionCountries.includes(country)
-                  ? getCountryDataBy('countryCode', country).countryName
-                  : getCountryDataBy(
+                <div
+                  className={
+                    isTaxDeductible
+                      ? styles.taxDeductibleCountry
+                      : styles.taxDeductibleCountryDisabled
+                  }
+                >
+                  {project.taxDeductionCountries.includes(country)
+                    ? getCountryDataBy('countryCode', country).countryName
+                    : getCountryDataBy(
                       'countryCode',
                       project.taxDeductionCountries[0]
                     ).countryName}
+                </div>
+                <div className={styles.downArrow}>
+                  <DownArrow color={isTaxDeductible ? '#87B738' : 'grey'} />
+                </div>
               </div>
-              <div className={styles.downArrow}>
-                <DownArrow color={isTaxDeductible ? '#87B738' : 'grey'} />
-              </div>
+            </React.Fragment>
+          )}
+
+          <div className={styles.horizontalLine} />
+
+          {paymentError && (
+            <div className={styles.paymentError}>{paymentError}</div>
+          )}
+          <div className={styles.finalTreeCount}>
+            <div className={styles.totalCost}>
+              {currency} {Sugar.Number.format(Number(treeCount * treeCost), 2)}
+              {/* {(treeCount * treeCost).toFixed(2)}{' '} */}
             </div>
-          </React.Fragment>
-        )}
-
-        <div className={styles.horizontalLine} />
-
-        {paymentError && (
-          <div className={styles.paymentError}>{paymentError}</div>
-        )}
-        <div className={styles.finalTreeCount}>
-          <div className={styles.totalCost}>
-            {currency} {Sugar.Number.format(Number(treeCount * treeCost), 2)}
-            {/* {(treeCount * treeCost).toFixed(2)}{' '} */}
+            <div className={styles.totalCostText}>
+              for {Sugar.Number.format(Number(treeCount))} Trees
           </div>
-          <div className={styles.totalCostText}>
-            for {Sugar.Number.format(Number(treeCount))} Trees
           </div>
+
+          {paymentSetup?.gateways?.stripe?.account &&
+            currency && (
+              <PaymentRequestCustomButton
+                country={country}
+                currency={currency}
+                amount={formatAmountForStripe(
+                  treeCost * treeCount,
+                  currency.toLowerCase()
+                )}
+                onPaymentFunction={onPaymentFunction}
+                continueNext={continueNext}
+              />
+            )}
+
         </div>
-
-        <div className={styles.actionButtonsContainer}>
-          <div style={{ width: '150px' }}>
-            {paymentSetup?.gateways?.stripe?.account &&
-              stripeAllowedCountries.includes(country) &&
-              currency && (
-                <PaymentRequestCustomButton
-                  country={country}
-                  currency={currency}
-                  amount={formatAmountForStripe(
-                    treeCost * treeCount,
-                    currency.toLowerCase()
-                  )}
-                  onPaymentFunction={onPaymentFunction}
-                />
-              )}
-
-            {/* {paymentRequest ? 'Or' : null} */}
-          </div>
-
-          <AnimatedButton
-            onClick={() => continueNext()}
-            className={styles.continueButton}
-          >
-            Continue
-          </AnimatedButton>
-        </div>
-      </div>
-      <SelectTaxDeductionCountryModal
-        openModal={openTaxDeductionModal}
-        handleModalClose={() => setOpenTaxDeductionModal(false)}
-        taxDeductionCountries={project.taxDeductionCountries}
-        setCountry={setCountry}
-        country={country}
-        setCurrency={setCurrency}
-        currency={currency}
-      />
-      <SelectCurrencyModal
-        openModal={openCurrencyModal}
-        handleModalClose={() => setOpenCurrencyModal(false)}
-        setCurrency={setCurrency}
-        currency={currency}
-        setCountry={setCountry}
-        country={country}
-      />
-    </>
-  );
+        <SelectTaxDeductionCountryModal
+          openModal={openTaxDeductionModal}
+          handleModalClose={() => setOpenTaxDeductionModal(false)}
+          taxDeductionCountries={project.taxDeductionCountries}
+          setCountry={setCountry}
+          country={country}
+          setCurrency={setCurrency}
+          currency={currency}
+        />
+        <SelectCurrencyModal
+          openModal={openCurrencyModal}
+          handleModalClose={() => setOpenCurrencyModal(false)}
+          setCurrency={setCurrency}
+          currency={currency}
+          setCountry={setCountry}
+          country={country}
+        />
+      </>
+    );
 }
 
 export default TreeDonation;
